Refactor

+++ api.py
@@ -27,14 +27,14 @@ Session = sessionmaker(bind=engine)
 session = Session()
 
 
-def barcode_check(barcode):
+def barcodeCheck(barcode):
     if barcodenumber.check_code("ean13", str(barcode)):
         return True
     return False
 
 
-def search_database(barcode):
-    if barcode_check(barcode):
+def searchDatabase(barcode):
+    if barcodeCheck(barcode):
         Session = sessionmaker(bind=engine)
         session = Session()
         query = session.query(Products).filter_by(barcode=barcode).first()
@@ -53,12 +53,12 @@ class Product:
         self.price = price
         self.quantity = quantity
 
-    def add_to_database(self):
+    def addToDatabase(self):
         Session = sessionmaker(bind=engine)
         session = Session()
         query = session.query(Products).filter_by(barcode=self.barcode).first()
         if query == None:
-            if barcode_check(self.barcode):
+            if barcodeCheck(self.barcode):
                 product = Products(barcode=self.barcode, name=self.name, price=self.price, quantity=self.quantity)
                 session.add(product)
                 session.commit()
@@ -67,11 +67,16 @@ class Product:
         session.close()
         return False
 
-def delete_products(barcode):
+
+
+
+def deleteProducts(barcode):
     try:
         Session = sessionmaker(bind=engine)
         session = Session()
         query = session.query(Products).filter_by(barcode=barcode).first()
+        if query == None:
+            return False
         session.delete(query)
         session.commit()
     except UnmappedInstanceError:
@@ -81,76 +86,112 @@ def delete_products(barcode):
     session.close()
     return True
 
-def rename_barcode(barcode, new_barcode):
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    query = session.query(Products).filter_by(barcode=barcode).first()
-    query.barcode = new_barcode
-    session.commit()
-    session.close()
 
-def rename_product(barcode, new_name):
+def renameBarcode(barcode, new_barcode):
+    if barcodeCheck(new_barcode):
+        Session = sessionmaker(bind=engine)
+        session = Session()
+        query = session.query(Products).filter_by(barcode=barcode).first()
+        if query == None:
+            return False
+        query.barcode = new_barcode
+        session.commit()
+        session.close()
+        return True
+    return False
+
+
+def renameProduct(barcode, new_name):
     Session = sessionmaker(bind=engine)
     session = Session()
     query = session.query(Products).filter_by(barcode=barcode).first()
+    if query == None:
+        return False
     query.name = new_name
     session.commit()
     session.close()
 
-def new_price(barcode, new_price):
+
+def newPrice(barcode, new_price):
     Session = sessionmaker(bind=engine)
     session = Session()
     query = session.query(Products).filter_by(barcode=barcode).first()
+    if query == None:
+        return False
     query.price = new_price
     session.commit()
     session.close()
 
-def check_stock(barcode):
+
+def checkStock(barcode):
     Session = sessionmaker(bind=engine)
     session = Session()
     query = session.query(Products).filter_by(barcode=barcode).first()
     session.close()
     if query == None:
         return False
-    data = {"quantity" : query.quantity}
+    data = {"quantity": query.quantity}
     return data
 
 
-def set_stock(barcode, amount):
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    query = session.query(Products).filter_by(barcode=barcode).first()
-    query.quantity = amount
-    session.commit()
-    session.close()
+def setStock(barcode, amount):
+    try:
+        Session = sessionmaker(bind=engine)
+        session = Session()
+        query = session.query(Products).filter_by(barcode=barcode).first()
+        if query == None:
+            return False
+        query.quantity = amount
+        session.commit()
+        session.close()
+    except:
+        return False
+    else:
+        return True
 
-def increase_stock(barcode, increase_amount):
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    query = session.query(Products).filter_by(barcode=barcode).first()
-    query.quantity = query.quantity + increase_amount
-    session.commit()
-    session.close()
 
-def decrease_stock(barcode, decreasing_amount):
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    query = session.query(Products).filter_by(barcode=barcode).first()
-    query.quantity = query.quantity - decreasing_amount
-    if query.quantity < 0:
-        query.quantity = 0
-    session.commit()
-    session.close()
+def increaseStock(barcode, increase_amount):
+    try:
+        Session = sessionmaker(bind=engine)
+        session = Session()
+        query = session.query(Products).filter_by(barcode=barcode).first()
+        if query == None:
+            return False
+        query.quantity = query.quantity + increase_amount
+        session.commit()
+        session.close()
+    except:
+        return False
+    else:
+        return True
+
+
+def decreaseStock(barcode, decreasing_amount):
+    try:
+        Session = sessionmaker(bind=engine)
+        session = Session()
+        query = session.query(Products).filter_by(barcode=barcode).first()
+        if query == None:
+            return False
+        query.quantity = query.quantity - decreasing_amount
+        if query.quantity < 0:
+            query.quantity = 0
+        session.commit()
+        session.close()
+    except:
+        return False
+    else:
+        return True
+
 
 barcodes = []
 names = ["Eggs", "Milk", "Bread", "Ice cream", "Flour", "Tea Bags", "Rat Poison", "Chocolate", "", ""]
-prices = [4.50, 3.25, 6.42, 10.00, 3.00, 2.40, 0.99, 4.30, 5.00,6.00]
+prices = [4.50, 3.25, 6.42, 10.00, 3.00, 2.40, 0.99, 4.30, 5.00, 6.00]
 
 
-#for i in range(10):
+# for i in range(10):
 #    
-p1 = Product(5901234123457, "Milk", 4.00, 3)
-
+# p1 = Product(5901234123457, "Milk", 4.00, 3)
 
 # print(search_database(5901234123457))
 
@@ -158,7 +199,7 @@ p1 = Product(5901234123457, "Milk", 4.00, 3)
 
 # print(p1.increase_stock(17))
 
-def product_get_all():
+def productGetAll():
     db = sqlite3.connect("database.db")
     c = db.cursor()
     c.execute("SELECT * FROM Products")
@@ -177,7 +218,7 @@ def product_get_all():
     return data
 
 
-def generate_barcode():
+def generateBarcode():
     pass
     # This is the new problem!
 
@@ -185,48 +226,79 @@ def generate_barcode():
 app = Flask(__name__)
 api = Api(app)
 
-product_put_args = reqparse.RequestParser()
-product_put_args.add_argument("barcode", type=int, help="Missing Barcode", required=True)
-product_put_args.add_argument("name", type=str, help="Missing Name", required=True)
-product_put_args.add_argument("price", type=float, help="Missing Price", required=True)
 
-class Home(Resource):
+class home(Resource):
     def get(self):
         return "Welcome to the shopping system Api"
 
-api.add_resource(Home, "/")
+
+api.add_resource(home, "/")
 
 
-class Product(Resource):
+class productSearch(Resource):
     def get(self, barcode):
-        return search_database(barcode=barcode)
+        return searchDatabase(barcode=barcode)
 
-api.add_resource(Product, "/product/<int:barcode>")
 
-class Product_add(Resource):
+api.add_resource(productSearch, "/product/<int:barcode>")
+
+product_post_args = reqparse.RequestParser()
+product_post_args.add_argument("barcode", type=int, help="Missing Barcode", required=True)
+product_post_args.add_argument("name", type=str, help="Missing Name", required=True)
+product_post_args.add_argument("price", type=float, help="Missing Price", required=True)
+product_post_args.add_argument("quantity", type=int, help="Missing Quantity", required=True)
+
+
+class productAdd(Resource):
     def post(self):
-        args = product_put_args.parse_args()
-        product = Product(args['barcode'], args['name'], args['price'], args['quantity'])
-        product.add_to_database()
-        
-api.add_resource(Product_add, "/product/add")
+        args = product_post_args.parse_args()
+        barcode = args['barcode']
+        name = args['name']
+        price = args['price']
+        quantity = args['quantity']
+
+        product = Product(barcode, name, price, quantity)
+        product.addToDatabase()
 
 
-class Product_Stock(Resource):
-    def get(self,  barcode):
-        return check_stock(barcode)
+api.add_resource(productAdd, "/product/add")
+
+product_stock_put_args = reqparse.RequestParser()
+product_stock_put_args.add_argument("quantity", type=int, help="Missing quantity", required=True)
+product_stock_put_args.add_argument("operation", type=str, help="Missing operation", required=True)
 
 
-api.add_resource(Product_Stock, "/product/stock/<int:barcode>")
+class productStock(Resource):
+    def get(self, barcode):
+        return checkStock(barcode)
+
+    def put(self, barcode):
+        args = product_stock_put_args.parse_args()
+        if args['operation'] == "set":
+            setStock(barcode, args['quantity'])
+        if args['operation'] == "increase":
+            increaseStock(barcode, args['quantity'])
+        if args['operation'] == "decrease":
+            decreaseStock(barcode, args['quantity'])
+        return productSearch(barcode)
 
 
-class Product_all(Resource):
+api.add_resource(productStock, "/product/stock/<int:barcode>")
+
+class productDelete(Resource):
+    def delete(self, barcode):
+        productDelete(barcode)
+
+api.add_resource(productDelete, "/product/delete/<int:barcode>")
+
+
+class productAll(Resource):
     def get(self):
-        return product_get_all()
+        return productGetAll()
 
 
-api.add_resource(Product_all, "/product/all")
+api.add_resource(productAll, "/product/all")
 
 if __name__ == "__main__":
-    app.run(host='0.0.0.0',port=8080,debug=True)  # For Replit
-    #app.run(debug=True)  # For Pycharm
+    # app.run(host='0.0.0.0',port=8080,debug=True)  # For Replit
+    app.run(debug=True)  # For Pycharm
